requirements.txt
IMDbPY==2021.4.18
omdb==0.10.1
python-dotenv==0.19.2
